[x/tour] content: explain shift operator in constants example
This addresses concerns raised in
https://go-review.googlesource.com/#/c/19325/

The literal number 1267650600228229401496703205376 is potentially
hard read. 1,267,650,600,228,229,401,496,703,205,376 was suggested
as an alternative.

The == operator has not yet been used in an example in the tour,
and so using it could potentially also confuse people.

Additionally it seems useful to hint at what << and >> mean, rather
than glossing over it entirely.

Change-Id: I89652eea138a91f3da5b7629b0edfaa96dcdbc23
Reviewed-on: https://go-review.googlesource.com/19370
Reviewed-by: Andrew Gerrand <adg@golang.org>
Reviewed-by: Rob Pike <r@golang.org>
Reviewed-by: Ralph Corderoy <ralph@inputplus.co.uk>
X-Tour-Commit: 7cfcba154e738d4ea2520693cd19204d1443861a