cmd/golangorg: use go/doc instead of x/website/content/doc
This is the first step towards deduplicating the doc folder so
that any changes in go/doc don't have to be backported to
the x/website repo. In the long term, the doc folder will
ultimately live in x/website, but that will be prioritized later
once x/tools has had website related code properly moved to
x/website.

Updates golang/go#29206

Change-Id: I1c5d200e9e72d3e24bc87f79e1d34faa8c9af36b
Reviewed-on: https://go-review.googlesource.com/c/160237
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>