cmd/golangorg: re-add documentation for flat presentation mode
The documentation of the "flat" presentation mode was removed
inadvertently as part of post-CLI-removal documentation cleanup
in CL 143077. The FlatDir flag still exists and is supported.
This change re-adds it to cmd/golangorg documentation.

Also fix a typo in the "src" presentation mode description.

(This is a re-apply of CL 162984 into x/website.)

Updates golang/go#25443
Updates golang/go#29206

Change-Id: I84145bf38f008678a8d8f3d3346f646aa7811886
Reviewed-on: https://go-review.googlesource.com/c/website/+/197721
Reviewed-by: Andrew Bonventre <andybons@golang.org>