cmd/golangorg: add GODOC environment variables to prod
While x/website relies on environment variables with the GOLANGORG_
prefix (as defined in app.prod.yaml), the required version of the
x/tools/godoc package relies on environment variables with the
GODOC_ prefix. So when it doesn't find them, it defaults to
GODOC_PROD and GODOC_ENFORCE_HOSTS being false. This causes the
website's share settings to not be set properly.

To resolve this issue we are adding GODOC_PROD and
GODOC_ENFORCE_HOSTS variables to x/website's app.prod.yaml
explicitly with the intention of deleting them in the future when
it is safe to do so.

Updates golang/go#29206

Change-Id: I5db74d680241b532f51bcc73eb51151e9f60483f
Reviewed-on: https://go-review.googlesource.com/c/164818
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>