[x/go.dev] learn: add data content and style for quickstart links
The quickstart section of the site will be links to third party learning
resources for the foreseeable future. This change adds a data file for
content like this, which should be useful for other pages that behave
similarly.

This change also includes layout for the quickstart portion of the learn
page, but without the gopher, as there are questions around the
breakpoint behaviour of it.
X-GoDev-Commit: ad5af2a896133497eb2a2d913833a2583137222f