cmd/golangorg: migrate from legacy to split health checks
For now, continue to use the existing /_ah/health HTTP endpoint
for both the liveness and readiness checks. This is simple and
close to the previous legacy health check behavior. In the future,
we can consider making the readiness check more sophisticated.

This change should make golangorg deployable again after the
deprecation of legacy health checks.¹

¹ https://cloud.google.com/appengine/docs/flexible/go/migrating-to-split-health-checks

Change-Id: Ia68d3b260dd3a5b66f969318a04b8f97639701d6
Reviewed-on: https://go-review.googlesource.com/c/website/+/201743
Reviewed-by: Katie Hockman <katie@golang.org>