[x/go.dev] themes/default: redirect linked pages
Some of our content pages are actually links to third party sites. If
someone attempts to visit the URL directly, we should redirect them to
the third party site.

The pages themselves do not appear in the sitemap, and shouldn't be
otherwise indexed or linked to anywhere.

Fixes b/143371871

Change-Id: Ie61e2fdcd4281ade2c0eec6345ebeeae535f3d6a
Reviewed-on: https://team-review.git.corp.google.com/c/golang/go.dev/+/590161
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Andrew Bonventre <andybons@google.com>
X-GoDev-Commit: a32be84ce5a13233c301ed9a470fb40793eca335