content: update README
The README does not mention that files that are newly added to
the content directory need to be added manually to static/gen.go
file before running go generate. This change fixes that.

Update the description of how the static content is used in concert
with the content from the main Go repository in more detail.

Replace go install && ./golangorg with go run, since it's shorter.

Add a TODO note mentioning improvements that can be made in the near
future to simplify this process.

Updates golang/go#29206

Change-Id: I8a85bcb783c7c10975aa6fd5a7ad464f05d42dd2
Reviewed-on: https://go-review.googlesource.com/c/website/+/207262
Reviewed-by: Andrew Bonventre <andybons@golang.org>