cmd/golangorg: update deploy instructions
Increase the separation between each step by giving them explicit
numbers. Expand on some of the deploy steps with more detail.

Update the default value of GO_REF to the current Go release branch
version. This way, it doesn't need to be overridden for each deploy.

Update references to the README.md file (it was recently renamed from
README.golangorg-app in CL 207438).

Change-Id: Ie34fed44a990a5c7ffe2acad68fd5c0f094f9143
Reviewed-on: https://go-review.googlesource.com/c/website/+/212239
Reviewed-by: Carlos Amedee <carlos@golang.org>