content/static/doc: copy doc/modules.md from GOROOT
Moved doc/modules.md from GOROOT to content/static/doc/modules.md.
Last updated in CL 214381. The corresponding deletion is CL 218240.

This documentation should be tied to the release, so it ought to be in
GOROOT. However, it won't be ready in time for Go 1.14, and we don't
want to wait until 1.15 to publish it.

Updates golang/go#33637

Change-Id: Ia7e59a74a764131cb1ae059b9f55a1dfc9fc9c9f
Reviewed-on: https://go-review.googlesource.com/c/website/+/218239
Reviewed-by: Bryan C. Mills <bcmills@google.com>