[x/go.dev] all: update CSP header to include golang.org as a connect-src
The site attempts to fetch the latest version of Go using a
CORS request to golang.org/dl/?mode=json, but doing so is currently
not permitted with the current CSP values. Update them to allow
connection to golang.org.

Also fixes retrieval logic to await the JSON response, since it
returns a promise, and properly sorts the versions.

Change-Id: If88bb3b12d097909e1f89c074b6726c12580470f
X-GoDev-Commit: 6e3d3fb4e7b9fb9e9f84235ce4283a8b43990a58