internal/dl: categorize new darwin/arm64 binaries better
We want the new darwin/arm64 installer and archive to be placed next
to the original darwin/amd64 files, not under "Other Ports". Also
highlight (make bold) the darwin/arm64 installer (but not archive).

The "pretty" version of the architecture will be displayed as "ARMv8",
and maybe should be made more macOS-specific in the future.

Updates golang/go#42756.

Change-Id: I5e95a3b93d30253d8fa291d153b3f96f49f9ab5d
Reviewed-on: https://go-review.googlesource.com/c/website/+/278787
Trust: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Alexander Rakoczy <alex@golang.org>