[x/go.dev] layouts/shortcodes: fix gopher.html template parse errors
This doesn't make sense, because each (...) is a bool,
so the first cannot be called as a function taking the second.

	{{if (eq $gopher "front") (eq $gopher "sticker1")}}

I am not sure whether the current web site works with some
older versions of Hugo, but it doesn't work on my machine
with the both v0.59.1 and the latest v0.81.0. It just looks like
it's always been broken, but then I'm confused about how the
web site itself is running.

In any event, this CL fixes the condition, which is clearly meant
to be an "or" of those two:

	{{if or (eq $gopher "front") (eq $gopher "sticker1")}}

But this is needlessly complex too, since eq will do that for us:

	{{if eq $gopher "front" "sticker1"}}

Fixed that condition and simplified a few others.

Now hugo runs successfully.

Change-Id: I02568c4e9331c71107488d44423535186e73e809
X-GoDev-Commit: 189cbf4f05b7bb855a3dc6f95c0acc57a2ba1434