_content/doc/tutorial: fix typo in add-a-test.html
In the "add-a-test" tutorial, the phrase "If this code" was used to
introduce a list of additions the user makes in the code. This
replaces "If" with "In".

Change-Id: I1343a648e15e2ffd4e49af2719f54aac6684e2ee
GitHub-Last-Rev: f7ae450460a4f1c93123dde1d541f3ba2debc83c
GitHub-Pull-Request: golang/website#31
Reviewed-on: https://go-review.googlesource.com/c/website/+/276772
Trust: Alberto Donizetti <alb.donizetti@gmail.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Keith Randall <khr@golang.org>