[x/go.dev] all: do not force every URL to end in slash
Hugo generates every page into its own directory with an index.html file,
so that all URLs end in slash. Our other web sites do not do this. For example:

	https://blog.golang.org/ports
	https://golang.org/pkg/math
	https://golang.org/ref/spec

As part of trying to simplify to a common core serving infrastructure,
drop the "every URL ends in slash" requirement for this site.

Note that due to what appears to be a bug in the App Engine config,
go.dev serves two of every page, both with and without the URL
For example, today on go.dev, these serve identical content
and neither redirects to the other:

	https://go.dev/about
	https://go.dev/about/

This CL is only changing the generated HTML links, not the locations
of the generated files. That will happen later.

Golden copies updated to match.

Change-Id: I5707360f6932878b4df689c593b28311c4cbc9b5
X-GoDev-Commit: b036875aad6369b69ce430db535c92df5d45eb6e