[x/go.dev] cmd/internal/site: switch to Goldmark
x/website uses Goldmark, so this brings go.dev in line with golang.org.

The only change to the golden output is a few automatically
generated heading IDs. It used to be that

	# Go's Concurrency

produced <h1 id="go-s-concurrency">.
Now it produces <h1 id="gos-concurrency">.

Similarly,

	# Go & You

used to produce <h1 id="go-you">.
Now it produces <h1 id="go--you">.

In theory this could lead to slightly broken links that
no longer target what they meant to target.
In practice it is doubtful that people have links to the
specific affected sections.

Change-Id: Ia0827e2f13fbd7f304983025dade694c50c49d6e
X-GoDev-Commit: 416036ecdf33ce51eeeb0d2bbd8ca98fd5fc7dc9