[x/go.dev] _content: correct various typographical problems for Goldmark
americanexpress:
The possessive of American Express is American Express's,
with a final s. The final s is only dropped for possessive plurals,
like the Smiths' house.

Goldmark also doesn't convert the trailing ' correctly,
so this makes the tree safer for Goldmark.

cloud:
The GoKit URL clearly does not end in [.

google:
The leading - looks like a single-element list in Markdown.
Change it to an em dash, which is more appropriate anyway.

Change-Id: I34beffaacef333ee7931c24b2b3503532d2e9de1
X-GoDev-Commit: fbb61a60c4e900e3151b46064bfd66b10041bad3