go.dev/cmd/frontend: fix learn.go.dev by redirecting to go.dev/learn
The old Hugo setup was strange in that Hugo thought it was generating
pages for go.dev/learn, and then we put in a redirect from go.dev/learn
to learn.go.dev, and then we had a whole separate app just for serving
files in that subtree. All the actual file serving was done in app.yaml using
static file directives, not anything in Go.

In the migration to a Go web server, I did not make learn.go.dev/ work,
so it was showing the main go.dev page instead, which is clearly bad.
I spent a little while trying to make the Go server behave like the old
setup, but in the end I gave up and removed all the subterfuge.
Now the pages are served from go.dev/learn, same as in the _content tree,
and there is a redirect from learn.go.dev/anything to go.dev/learn/anything.
Given how much trouble I had getting the server to work the other way,
it seemed like we'd keep tripping over this special case as we make more changes.
And given that the long-term plan is to move everything onto go.dev,
it seemed better all around to just start telling the truth now.

This fix is already deployed, to fix the broken learn.go.dev.

Change-Id: I774850ade327623eefffc785f923f5002f3abff8
Reviewed-on: https://go-review.googlesource.com/c/website/+/322970
Trust: Russ Cox <rsc@golang.org>
Run-TryBot: Russ Cox <rsc@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>