go.dev/_content/css: various cleanups
1. Remove all letter-spacing overrides

Overriding the letter-spacing might have made sense when we
were hard-coding the font (I have my doubts), but it certainly
does not make sense now that we are defaulting to the system font
instead of a hard-coded font.

The effect of the letter-spacing now is to override the font's
kerning tables (as if we could do better!) and gives the site a
bit of an alien look. The only time a letter-spacing override
really works is in ALL CAPS headlines, but we don't have any.

2. Remove font-weight override for article headings

The “large but not bold” headings are another thing that don't
really work consistently in the variety of fonts users might see.
Go back to headings being bold by removing the overrides.

Also rewrite a few font-weight: 400 to font-weight: normal for consistency.

3. Change body text to darker grey

Using 3e4042 for body text is too light and hurts readability.
pkg.go.dev uses 202224 for body text; use that here too.
Global search and replace to keep the same number of colors in our palette.

4. Regularize page max-width, line-height

Solutions articles were allowed to be 75.75rem wide,
but regular articles were capped at 45rem, which is
far too narrow when you're using a large screen.
Let all articles expand to 75.75rem.
(Not sure what is magic about 75.75 but it's used elsewhere too.)

Regular line-height to a default of 1.4 for the whole page.
Rutter's book _Web Typography_ says 1.4 is a good default
for a variety of fonts. The blog was already overriding to 1.4
because that was most readable for the long articles.
Many other elements were using 1.5rem, which can then be deleted.
Article lists were set to 1.45, so delete that too in favor of consistency.
Article text was set to 1.75rem, which is far too wide and hurts
readability (the blog was overriding this specific setting),
so drop that too.

5. Narrow spaces in code font

This rule will let us stop insisting that people write things like

	Use <code>go</code> <code>get</code> to ...

(See for example doc/go1.*.html.)

Change-Id: I2154b03296ae1d56bc9bc416a1e3fb9e930d7d9c
Reviewed-on: https://go-review.googlesource.com/c/website/+/362015
Trust: Russ Cox <rsc@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Website-Publish: Russ Cox <rsc@golang.org>