go.dev/_content/js: delete h2 manipulation in misc.js
It's unclear what “correct scroll” means here, but it's definitely
incorrect to blow away any HTML (such as links) in every h2
and replace it with the text-only representation of the original h2.
Delete that code. If we figure out what this means, we can look
for a better way to solve this problem (probably scroll-margin-top?).

Also, make it clearer that the stickyNav setup is a no-op without a sticky nav.

Preparation for the golang.org -> go.dev move.

Change-Id: Ib8930a4e3aba688d645f689d3c2ee90b3e09229f
Reviewed-on: https://go-review.googlesource.com/c/website/+/362499
Trust: Russ Cox <rsc@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>